ADC128D818 library is connected to the software
The ADC128D818 library is updated in order to:
+ clear all non necessaries sentences
+ to initialize only once the I2C driver (speed and pins)
+ add delays time required by the chip (this is a preliminary implementation and must be refactored)

It is also modified the ADC128D818 library initialization to set the chip in continuous mode and enabling all channel (for both chips).

Raw to engineering values conversion is adapted for the ADC128D818 values read from the input mapped to rely's voltage and current measure.

It is observed that the ADC128D818 library could not be initialized twice for the same chip. After the start bit is set (last instruction of the initialization), the busy status register is always set to one. This must be fixed because a software reset could cause a wrong ADC128D818 chip initialization.