Two bugs fixed in ADC128D818 library
The firs bug is related to bits interpretation from status register. The bits definition was wrong. Also fixed the way of checking the registers value during the chip initialization.
The second bug is related to the rate configuration. The "one shot" conversion rate configuration was wrong. According to chip datasheet, for the one shot conversion the "One shot register" must be set. This was fixed. It is also added a register setting to "set chip in shutdown mode" before modifying the rate register (it is needed to change to continuous mode according to datasheet).
Also a bug is fixed when ADC channels are sensed. There are only 2 ADC (ID 0 and ID 1), and due to a bug the software was sensing 3 ADC (ID 0 to 2).